refactor(daemon): extract hooks config loading
